feat: implement profile updater
