+++ src/eli/profile.py
@@ -5,19 +5,6 @@ An Eli profile used to configure Eli behaviour for a specific use-case.
 
 from configurable import Configurable
 
-
-class NoProfile(Exception):
-    """A profile was not found."""
-
-
-class NoProfileSetting(Exception):
-    """A profile setting was not found."""
-
-
-class ProfileAlreadyExists(Exception):
-    """A profile already exists with that name."""
-
-
 class Profile(Configurable):
     """The base class for a Profile."""
 
@@ -39,40 +26,3 @@ class Profile(Configurable):
         )
 
 
-class _Profiles(Configurable):
-    """manage a collection of profiles."""
-
-    _DEFAULT_PROFILES = [
-        Profile(name="Eli", debug=False, cls_on_submit=True, model=""),
-        Profile(name="Debugger", debug=True, cls_on_submit=True, model=""),
-    ]
-
-    def __init__(self, profiles):
-        self.loaded = {}
-        self.load(self._DEFAULT_PROFILES)
-        self.load(profiles)
-
-    def load(self, profiles: [Profile]):
-        try:
-            for profile in profiles:
-                name = profile.name
-                self.loaded.update({name: profile})
-        except:
-            raise ProfileAlreadyExists(
-                f"A profile called {profile.name} already exists."
-            )
-
-    def get_profile(self, name: str):
-        try:
-            return self.loaded[name]
-        except:
-            raise NoProfile(f"Profile not found: {name}.")
-
-    def get_profile_setting(self, name: str, setting: str):
-        profile = self.get_profile(name)
-        try:
-            return profile[setting]
-        except:
-            raise NoProfileSetting(
-                f"Profile {name} does not have the setting: {setting}"
-            )
