--- src/eli/ui/__init__.py
@@ -1,1 +1,0 @@
-from .ui import Ui
